adding post controller
